passkey_abdication_backend/pyproject.toml
[project]
name = "pad"
version = "0.1.0"
description = ""
authors = [
    {name = "Atharva Chincholkar",email = "atharva@pureid.io"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
django-split-settings = "^1.3.1"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
django-cors-headers = "^4.3.1"
pytz = "^2024.1"
phonenumbers = "^8.13.37"
nanoid = "^2.0.0"
django-filter = "^24.3"
django-ipware = "^7.0.1"
selenium = "^4.31.0"
undetected-chromedriver = "^3.5.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"
django-extensions = "^3.2.3"
django-silk = "^5.2.0"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
